requirements.txt
absl-py==2.3.0
astunparse==1.6.3
certifi==2025.4.26
charset-normalizer==3.4.2
customtkinter==5.2.2
darkdetect==0.8.0
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.72.1
h5py==3.13.0
idna==3.10
imageai==3.0.3
ImageHash==4.3.2
keras==3.10.0
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
namex==0.1.0
numpy==2.1.3
opt_einsum==3.4.0
optree==0.16.0
packaging==25.0
pillow==11.2.1
pillow_heif==0.22.0
protobuf==5.29.5
Pygments==2.19.1
PyWavelets==1.8.0
requests==2.32.3
rich==14.0.0
scipy==1.15.3
six==1.17.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
typing_extensions==4.14.0
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
